requirements.txt
# Core LangChain
langchain==0.3.7
langchain-openai==0.2.9
langchain-chroma==0.1.4
langchain-community==0.3.7

# Vector DB
chromadb==0.5.18

# Document parsing
pymupdf==1.24.12
python-docx==1.1.2
docx2txt==0.8

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# UI
streamlit==1.40.1

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
tiktoken==0.8.0